feat: added react router and renamed components
